hwc: Don't lock eventControl for vsync events
Locking here can lead to a delay in vsync disable/enable
where prepare/set is not yet completed. The HWC spec says
that eventControl needs to take effect immediately.
This lock was earlier present when we waited for
SF to unblank primary at startup and the vsync control
ioctls could fail. This is no longer necessary as we unblank
primary on startup now.

Bug: 9669813
Change-Id: I2bf18ae9235f8dd428a495a9a23af7f00f34ff1d
Signed-off-by: Iliyan Malchev <malchev@google.com>